recompute p2g in mdl step to ensure the score is distance between CRE and TSS. Checked downloading annotations

+++ workflow/scripts/methods/dictys/mdl.py
@@ -14,6 +14,8 @@ parser.add_argument('-p', '--path_out', required=True)
 parser.add_argument('-w', '--working_dir', required=True)
 parser.add_argument('-v', '--device', required=True)
 parser.add_argument('-t', '--threads', required=True)
+parser.add_argument('-e', '--ext', required=True)
+parser.add_argument('-a', '--gene_annotation', required=True)
 
 args = vars(parser.parse_args())
 mudata = args['path_data']
@@ -21,8 +23,10 @@ p2g_data = args['p2g_data']
 tfb_data = args['tfb_data']
 path_out = args['path_out']
 working_dir = args['working_dir']
-device = f'cuda:{args['device']}'
+device = f'cuda:{args["device"]}'
 threads = int(args['threads'])
+distance = int(args['ext'])
+annot = args['gene_annotation']
 
 
 # QC Expression matrix
@@ -31,18 +35,23 @@ proc_rna_filename = os.path.join(working_dir, "expression_proc.tsv.gz")
 
 data = md.read(mudata)
 rna_X = pd.DataFrame(np.array(data['rna'].layers['counts'].todense()).T, columns=data['rna'].obs.index, index=data['rna'].var.index)
-rna_X.to_csv(rna_filename, sep="\t", compression="gzip") 
+if not os.path.exists(rna_filename):
+    rna_X.to_csv(rna_filename, sep="\t", compression="gzip") 
 dictys.preproc.qc_reads(rna_filename, proc_rna_filename, 50, 10, 0, 0, 0, 0)  # all cell filters are turned off
 rna_X_proc = pd.read_csv(proc_rna_filename, sep='\t', index_col=0)
 genes = rna_X_proc.index.to_numpy()
 
 
-# Subset Peak-Gene linking to preserve non-spurious genes
-p2g = pd.read_csv(p2g_data)
-p2g.columns = ['region', 'target', 'dist']  # What if the method output scores where sign matters, unlike distance?
-p2g = p2g[p2g.target.apply(lambda x: x in genes)]
-p2g_name = os.path.join(working_dir, 'p2g_reformat.tsv.gz')
-p2g.to_csv(p2g_name, sep='\t', index=False)
+# Write ATAC to working directory
+atac_filename = os.path.join(working_dir, "atac_peak.tsv.gz")
+atac_peak_names = [n.replace('-', ':') for n in data['atac'].var.index]
+atac_X = pd.DataFrame(np.zeros((data['atac'].var.index.shape[0], 1)), index=atac_peak_names, columns=['placeholder'])
+if not os.path.exists(atac_filename):
+    atac_X.to_csv(atac_filename, sep="\t", compression="gzip") 
+
+# Identify all peaks that are within Xbp of annotated TSS 
+dist_filename = os.path.join(working_dir, "tssdist.tsv.gz")
+os.system(f'python3 -m dictys chromatin tssdist --cut {distance//2} {proc_rna_filename} {atac_filename} {annot} {dist_filename}')
 
 
 # Subset TF-Peak linking to preserve non-spurious TFs
@@ -56,7 +65,7 @@ tfb[['TF', 'loc', 'score']].to_csv(tfb_name, sep='\t', index=False)
 # Compute TF-Gene linking from TF-Peak and Peak-Gene linkings
 tfg_name = os.path.join(working_dir, 'linking.tsv.gz')
 mask_name = os.path.join(working_dir, 'binlinking.tsv.gz')
-os.system(f'python3 -m dictys chromatin linking {tfb_name} {p2g_name} {tfg_name}')
+os.system(f'python3 -m dictys chromatin linking {tfb_name} {dist_filename} {tfg_name}')
 os.system(f'python3 -m dictys chromatin binlinking {tfg_name} {mask_name} 20')
